refactor(security): Restricting access to certains data

+++ Application/SmartWeather.Api/Controllers/Station/StationController.cs
@@ -7,7 +7,11 @@ using SmartWeather.Services.Stations;
 using SmartWeather.Api.Controllers.Station.Dtos;
 using SmartWeather.Api.Controllers.Station.Dtos.Converters;
 using SmartWeather.Entities.Station;
-
+using SmartWeather.Services.Authentication;
+using SmartWeather.Entities.User;
+using SmartWeather.Services.Components;
+using Microsoft.IdentityModel.Tokens;
+using SmartWeather.Api.Controllers.MeasurePoint.Dtos;
 
 [Route("api/[controller]")]
 [ApiController]
@@ -15,10 +19,12 @@ using SmartWeather.Entities.Station;
 public class StationController : ControllerBase
 {
     private readonly StationService _stationService;
+    private readonly AuthenticationService _authenticationService;
 
-    public StationController(StationService stationService)
+    public StationController(StationService stationService, AuthenticationService authenticationService)
     {
         _stationService = stationService;
+        _authenticationService = authenticationService;
     }
 
     [HttpPost(nameof(Create))]
@@ -27,7 +33,7 @@ public class StationController : ControllerBase
         ApiResponse<StationResponse> response;
         StationResponse formattedResponse;
 
-        if (String.IsNullOrWhiteSpace(request.Name) ||
+        if (string.IsNullOrWhiteSpace(request.Name) ||
             !(request.UserId > 0))
         {
             return BadRequest(ApiResponse<StationResponse>.Failure(BaseResponses.ARGUMENT_ERROR));
@@ -35,13 +41,64 @@ public class StationController : ControllerBase
 
         try
         {
-            Entities.Station.Station createdStation = _stationService.AddNewStation(request.Name, request.MacAddress, request.Latitude, request.Longitude, request.UserId);
+            Station createdStation = _stationService.AddNewStation(request.Name, request.MacAddress, request.Latitude, request.Longitude, request.UserId);
             formattedResponse = StationResponseConverter.ConvertStationToStationResponse(createdStation);
             response = ApiResponse<StationResponse>.Success(formattedResponse);
         }
         catch (Exception ex)
         {
-            response = ApiResponse<StationResponse>.Failure(String.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            response = ApiResponse<StationResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return BadRequest(response);
+        }
+
+        return Ok(response);
+    }
+
+    [HttpPost(nameof(Claim))]
+    public ActionResult<ApiResponse<StationResponse>> Claim(StationClaimRequest request)
+    {
+        ApiResponse<StationResponse> response;
+        StationResponse formattedResponse;
+
+        var token = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", "");
+
+
+        if (string.IsNullOrWhiteSpace(request.MacAddress) ||
+            string.IsNullOrEmpty(token))
+        {
+            return BadRequest(ApiResponse<StationResponse>.Failure(BaseResponses.ARGUMENT_ERROR));
+        }
+
+        Console.WriteLine(token);
+
+        try
+        {
+            int userId = _authenticationService.GetUserIdFromToken(token);
+
+            if (userId <= 0)
+            {
+                throw new Exception("Unable to retreive user info from token");
+            }
+
+            Station? retreivedStation = _stationService.GetStationByMacAddress(request.MacAddress);
+            if (retreivedStation == null)
+            {
+                throw new Exception("MacAdress doesn't correspond to any registered station");
+            }
+
+            retreivedStation = _stationService.UpdateStation(retreivedStation.Id,
+                                          retreivedStation.Name,
+                                          retreivedStation.MacAddress,
+                                          retreivedStation.Latitude,
+                                          retreivedStation.Longitude,
+                                          userId);
+
+            formattedResponse = StationResponseConverter.ConvertStationToStationResponse(retreivedStation);
+            response = ApiResponse<StationResponse>.Success(formattedResponse);
+        }
+        catch (Exception ex)
+        {
+            response = ApiResponse<StationResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
             return BadRequest(response);
         }
 
@@ -51,11 +108,13 @@ public class StationController : ControllerBase
     [HttpPut(nameof(Update))]
     public ActionResult<ApiResponse<StationResponse>> Update(StationUpdateRequest request)
     {
+        var token = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", "");
+
         ApiResponse<StationResponse> response;
         StationResponse formattedResponse;
 
         if (!(request.Id > 0) ||
-            String.IsNullOrWhiteSpace(request.Name) ||
+            string.IsNullOrWhiteSpace(request.Name) ||
             !(request.Id > 0))
         {
             return BadRequest(ApiResponse<StationResponse>.Failure(BaseResponses.ARGUMENT_ERROR));
@@ -63,7 +122,16 @@ public class StationController : ControllerBase
 
         try
         {
-            Entities.Station.Station updatedStation = _stationService.UpdateStation(request.Id, request.Name, request.MacAddress, request.Latitude, request.Longitude, request.UserId);
+            int userId = _authenticationService.GetUserIdFromToken(token);
+            Role userRole = _authenticationService.GetUserRoleFromToken(token);
+
+            if (!_stationService.IsOwnerOfStation(userId, request.Id) &&
+                !RoleAccess.ADMINISTRATORS.Contains(userRole))
+            {
+                throw new UnauthorizedAccessException("Your are not allowed to update this information");
+            }
+
+            Station updatedStation = _stationService.UpdateStation(request.Id, request.Name, request.MacAddress, request.Latitude, request.Longitude, request.UserId);
             formattedResponse = StationResponseConverter.ConvertStationToStationResponse(updatedStation);
             response = ApiResponse<StationResponse>.Success(formattedResponse);
         }
@@ -79,7 +147,8 @@ public class StationController : ControllerBase
     [HttpDelete(nameof(Delete))]
     public ActionResult<ApiResponse<EmptyResponse>> Delete(int idStation)
     {
-        // Later will need to restrict this to admin or current token user privilege
+        var token = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", "");
+
         ApiResponse<EmptyResponse> response;
 
         if (!(idStation > 0))
@@ -89,6 +158,15 @@ public class StationController : ControllerBase
 
         try
         {
+            int userId = _authenticationService.GetUserIdFromToken(token);
+            Role userRole = _authenticationService.GetUserRoleFromToken(token);
+
+            if (!_stationService.IsOwnerOfStation(userId, idStation) &&
+                !RoleAccess.ADMINISTRATORS.Contains(userRole))
+            {
+                throw new UnauthorizedAccessException("Your are not allowed to delete this information");
+            }
+
             bool isStationDeleted = _stationService.DeleteStation(idStation);
             if (isStationDeleted)
             {
@@ -112,7 +190,8 @@ public class StationController : ControllerBase
     [HttpGet(nameof(GetFromUser))]
     public ActionResult<ApiResponse<StationListResponse>> GetFromUser(int userId)
     {
-        // Need to implement
+        var token = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", "");
+
         ApiResponse<StationListResponse> response;
 
         if (!(userId > 0))
@@ -122,6 +201,15 @@ public class StationController : ControllerBase
 
         try
         {
+            int userIdToken = _authenticationService.GetUserIdFromToken(token);
+            Role userRole = _authenticationService.GetUserRoleFromToken(token);
+
+            if (userIdToken != userId &&
+                !RoleAccess.GLOBAL_READING_ACCESS.Contains(userRole))
+            {
+                throw new UnauthorizedAccessException("Your are not allowed to get this information");
+            }
+
             IEnumerable<Station> stationList = _stationService.GetFromUser(userId);
             StationListResponse formattedResponse = StationListResponseConverter.ConvertStationListToStationListResponse(stationList);
             if (stationList.Any())
@@ -134,9 +222,19 @@ public class StationController : ControllerBase
             }
 
         }
+        catch (SecurityTokenException ex)
+        {
+            response = ApiResponse<StationListResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return Unauthorized(response);
+        }
+        catch (UnauthorizedAccessException ex)
+        {
+            response = ApiResponse<StationListResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return Unauthorized(response);
+        }
         catch (Exception ex)
         {
-            response = ApiResponse<StationListResponse>.Failure(String.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            response = ApiResponse<StationListResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
             return BadRequest(response);
         }
 
@@ -147,7 +245,8 @@ public class StationController : ControllerBase
     [HttpGet(nameof(GetById))]
     public ActionResult<ApiResponse<StationResponse>> GetById(int idStation)
     {
-        // Later will need to restrict this to admin or current token user privilege
+        var token = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", "");
+        
         ApiResponse<StationResponse> response;
 
         if (!(idStation > 0))
@@ -157,6 +256,15 @@ public class StationController : ControllerBase
 
         try
         {
+            int userId = _authenticationService.GetUserIdFromToken(token);
+            Role userRole = _authenticationService.GetUserRoleFromToken(token);
+
+            if (!_stationService.IsOwnerOfStation(userId, idStation) &&
+                !RoleAccess.GLOBAL_READING_ACCESS.Contains(userRole))
+            {
+                throw new UnauthorizedAccessException("Your are not allowed to get this information");
+            }
+
             Station stationRetreived = _stationService.GetStationById(idStation);
             if (stationRetreived != null)
             {
@@ -168,33 +276,62 @@ public class StationController : ControllerBase
             }
 
         }
+        catch (SecurityTokenException ex)
+        {
+            response = ApiResponse<StationResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return Unauthorized(response);
+        }
+        catch (UnauthorizedAccessException ex)
+        {
+            response = ApiResponse<StationResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return Unauthorized(response);
+        }
         catch (Exception ex)
         {
-            response = ApiResponse<StationResponse>.Failure(String.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            response = ApiResponse<StationResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
             return BadRequest(response);
         }
 
         return Ok(response);
     }
 
-    //[HttpGet(nameof(GetAll))]
-    //public ActionResult<ApiResponse<UserListResponse>> GetAll()
-    //{
-    //    // Later will need to restrict this to admin or current token user privilege
-    //    ApiResponse<UserListResponse> response;
+    [HttpGet(nameof(GetAll))]
+    public ActionResult<ApiResponse<StationListResponse>> GetAll()
+    {
+        var token = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", "");
+        
+        ApiResponse<StationListResponse> response;
 
-    //    try
-    //    {
-    //        IEnumerable<SmartWeather.Entities.User.User> usersRetreived = _userService.GetUserList(null);
-    //        response = ApiResponse<UserListResponse>.Success(UserResponseConverter.ConvertUserListToUserListResponse(usersRetreived));
-    //    }
-    //    catch (Exception ex)
-    //    {
-    //        response = ApiResponse<UserListResponse>.Failure(String.Format(BaseResponses.FORMAT_ERROR, ex.Message));
-    //        return BadRequest(response);
-    //    }
+        try
+        {
+            Role userRole = _authenticationService.GetUserRoleFromToken(token);
 
-    //    return Ok(response);
-    //}
+            if (!RoleAccess.ADMINISTRATORS.Contains(userRole))
+            {
+                throw new UnauthorizedAccessException("Your are not allowed to get this information");
+            }
+
+            IEnumerable<Station> stationsRetreived = _stationService.GetAll();
+            response = ApiResponse<StationListResponse>.Success(StationListResponseConverter.ConvertStationListToStationListResponse(stationsRetreived));
+        
+        }
+        catch (SecurityTokenException ex)
+        {
+            response = ApiResponse<StationListResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return Unauthorized(response);
+        }
+        catch (UnauthorizedAccessException ex)
+        {
+            response = ApiResponse<StationListResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return Unauthorized(response);
+        }
+        catch (Exception ex)
+        {
+            response = ApiResponse<StationListResponse>.Failure(string.Format(BaseResponses.FORMAT_ERROR, ex.Message));
+            return BadRequest(response);
+        }
+
+        return Ok(response);
+    }
 
 }
